pyproject.toml
[project]
name = "dag-as-config"
version = "0.1.0"
description = "Apache Airflow DAG generator from config yaml"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "airflow-provider-kafka>=0.2.5",
    "apache-airflow[snowflake]>=3.1.3",
    "apache-airflow-providers-amazon>=9.17.0",
    "apache-airflow-providers-snowflake>=6.7.0",
    "boto3>=1.42.0",
    "dag-factory>=1.0.1",
    "pandas>=2.3.3",
]

[tool.pytest.ini_options]
pythonpath = "dags"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
